backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0